Added support for Seahorse shapes bound to the http response status code.
The Glacier#get_job_output operation specifies a `status` member
in the output shape.  This member specifies a location of
"http_status", but this was not correctly managed by the shape
class when grouping members by location.

The InputShape and OutputShape classes now provide a http_status_member.
This could be cleaned up a bit and moved out of the base class
as only OutputShape objects can actually bind members to
the http_status.

Fixes #34.